==> main/05a_species_area_nonpooled.Rmd <==
---
title: "species_area_nonpooled"
output: html_document
date: "2024-06-13"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(dplyr)
library(ggplot2)

library(MetBrewer)
library(wesanderson)
library(patchwork)
library(pals)
```


```{r}
fung_data <- readRDS("../intermediates/culled_fung_asv_table_decontamed_nonpooled.rds")
bact_data <- readRDS("../intermediates/culled_bact_asv_table_decontamed_nonpooled.rds")

fung_rep_dat <- readRDS("../intermediates/fung_replication_data_v2.rds")
fung_rep_dat <- fung_rep_dat %>% group_by(bro_and_comp) %>% mutate(newnum = n()) %>% ungroup

bact_rep_dat <- readRDS("../intermediates/bact_replication_data_v2.rds")
bact_rep_dat <- bact_rep_dat %>% group_by(bro_and_comp) %>% mutate(newnum = n()) %>% ungroup

bro_metadata <- read.csv("../metadata/sar_metadata_clean - Sheet1.csv")
```

```{r}
brostuff <- bro_metadata %>% tidyr::separate(Label, c("Project", "bro_num", "num"))
brostuff$bro_num <- paste0(brostuff$bro_num, brostuff$num)

fung_rep_dat$water_volume <- brostuff$Water.volume...in.lab...mL.[match(fung_rep_dat$Bro1, brostuff$bro_num)]
fung_rep_dat$detritus_wt <- brostuff$Dry.weight.detritus[match(fung_rep_dat$Bro1, brostuff$bro_num)]

bact_rep_dat$water_volume <- brostuff$Water.volume...in.lab...mL.[match(bact_rep_dat$Bro1, brostuff$bro_num)]
bact_rep_dat$detritus_wt <- brostuff$Dry.weight.detritus[match(bact_rep_dat$Bro1, brostuff$bro_num)]
```

```{r}
fung_rep_dat$newCompartment[fung_rep_dat$Compartment=="W02"] = "W02"
fung_rep_dat$newCompartment[fung_rep_dat$Compartment=="WPool"] = "WPool"

bact_rep_dat$newCompartment[bact_rep_dat$Compartment=="W02"] = "W02"
bact_rep_dat$newCompartment[bact_rep_dat$Compartment=="WPool"] = "WPool"

# remove Wpool samples in bact?
bact_rep_dat <- bact_rep_dat[!(grepl("WPool", bact_rep_dat$sample)),]

bact_data <- bact_data[,colnames(bact_data) %in% bact_rep_dat$sample,]
bact_data <- bact_data[rowSums(bact_data)>0,]
```

### Test check - regress replicate number and water volume or detritus weight

```{r}
# fff <- fung_rep_dat %>% group_by(bro_and_comp, newnum, water_volume, detritus_wt) %>% summarise()
# bbb <- bact_rep_dat %>% group_by(bro_and_comp, newnum, water_volume, detritus_wt) %>% summarise()
# 
# 
# fwv <- lm(fff$newnum ~ fff$water_volume); summary(fwv)
# bwv <- lm(bbb$newnum ~ bbb$water_volume); summary(bwv)
# 
# fdw <- lm(fff$newnum ~ fff$detritus_wt); summary(fdw)
# bdw <- lm(bbb$newnum ~ bbb$detritus_wt); summary(bdw)
```


Functions!
```{r}
get_data_for_regress <- function(rep_dat, compartment) {
  filt_dat <- rep_dat[grepl(compartment, rep_dat$newCompartment),]
  #filt_dat <- rep_dat[rep_dat$newCompartment == compartment,]
  
  filt_dat$log_richness <- log(filt_dat$richness)
  # 
  # dat_summ <- filt_dat %>% group_by(Bro1) %>% summarize(log_rich = mean(log_richness))
  # 
  # dat_summ_se <- filt_dat %>% group_by(Bro1) %>% 
  #   summarize(se = sd(log_richness)/sqrt(length(log_richness)))
  # 
  # dat_summ$se <- dat_summ_se$se
  # 
  # dat_summ$water_volume <- filt_dat$water_volume[match(dat_summ$Bro1, filt_dat$Bro1)]
  # dat_summ$detritus_wt <- filt_dat$detritus_wt[match(dat_summ$Bro1, filt_dat$Bro1)]
  # 
  # res <- list()
  # 
  # res[["full_data"]] <- filt_dat
  # res[["summarized_data"]] <- dat_summ
  # 
  # return(res)
  return(filt_dat)
}


fung_inv_regress_dat <- get_data_for_regress(rep_dat = fung_rep_dat,
                                             compartment = "Inverts")

fung_det_regress_dat <- get_data_for_regress(rep_dat = fung_rep_dat,
                                             compartment = "Detritus")

fung_water_regress_dat <- get_data_for_regress(rep_dat = fung_rep_dat,
                                             compartment = "W")



bact_inv_regress_dat <- get_data_for_regress(rep_dat = bact_rep_dat,
                                             compartment = "Inverts")

bact_det_regress_dat <- get_data_for_regress(rep_dat = bact_rep_dat,
                                             compartment = "Detritus")

bact_water_regress_dat <- get_data_for_regress(rep_dat = bact_rep_dat,
                                             compartment = "W02")

# bact_wpool_regress_dat <- get_data_for_regress(rep_dat = bact_rep_dat,
#                                              compartment = "WPool")

```


```{r}
# bootstrap function
bfun <- \(x, y, R=1000, seed=134) {
  FUN <- \() {
    i <- sample(seq_len(length(x)), replace=TRUE)
    y <- y[i]
    x <- x[i]
    return(lm(y ~ x)$coefficients)

    #return()
  }
  set.seed(seed)
  r <- t(replicate(R, FUN()))
  bootdist <- r[, 2]

  return(bootdist)

}

```

Regressions and bootstraps
```{r}
# volume = detritus weight
fung_regress_data_list = list(fung_inv_regress_dat,
                              fung_det_regress_dat,
                              fung_water_regress_dat)

fung_detwt_boot_slopes <- lapply(fung_regress_data_list,
                                 function(data) bfun(x = log(data$detritus_wt), y = data$log_richness, R = 10000, seed=1899))


fung_regular_lms_detritus_wt <- lapply(fung_regress_data_list,
                           function(x) lm(log_richness ~ log(detritus_wt), data = x))

fung_reg_lm_slopes_detritus_wt <- lapply(fung_regular_lms_detritus_wt, function(y) coef(y)[2])

fung_tts_detritus_wt <- list()

for (i in 1:length(fung_regress_data_list)) {
  fung_tts_detritus_wt[[i]] = t.test(unlist(fung_detwt_boot_slopes[[i]]), mu = fung_reg_lm_slopes_detritus_wt[[i]])
}


##### water volume
fung_wvol_boot_slopes <- lapply(fung_regress_data_list,
                                 function(data) bfun(x = log(data$water_volume), y = data$log_richness, R = 10000, seed=222))

fung_regular_lms_water_volume <- lapply(fung_regress_data_list,
                           function(x) lm(log_richness ~ log(water_volume), data = x))

fung_reg_lm_slopes_water_volume <- lapply(fung_regular_lms_water_volume, function(y) coef(y)[2])

fung_tts_water_volume <- list()

for (i in 1:length(fung_regress_data_list)) {
  fung_tts_water_volume[[i]] = t.test(unlist(fung_wvol_boot_slopes[[i]]), mu = fung_reg_lm_slopes_water_volume[[i]])
}
```

Air
```{r}
# air
# richness per sample vs water volume

# We left the sampler out in the field to collect samples of the air microbiome above the bromeliad garden for 1 hour, 2.5 hours, 4.5 hours, and 6 hours with n=6 replicates per time point. These time points correspond to 990, 2475, 4455, and 5940 liters respectively. 

fung_air <- fung_rep_dat[fung_rep_dat$newCompartment=="Air",]
fung_air <- subset(fung_air, select = -c(detritus_wt, water_volume))
fung_air$combo <- paste(fung_air$Bro1, fung_air$Compartment, sep = "_")

fung_air<- fung_air %>%
  mutate(air_volume = case_when(
    Compartment == "T1" ~ 990,
    Compartment == "T2" ~ 2475,
    Compartment == "T3" ~ 4455,
    Compartment == "T4" ~ 5940
  ))

fung_air$log_richness <- log(fung_air$richness)

##### bootstrap and regressions
fung_air_boot_slopes <- bfun(x = log(fung_air$air_volume), y = fung_air$log_richness, R = 10000, seed=111)

fung_regular_lm_air_volume <- lm(log_richness ~ log(air_volume), data = fung_air)

fung_reg_lm_slope_air_volume <- coef(fung_regular_lm_air_volume)[2]

fung_tts_air_volume <- t.test(unlist(fung_air_boot_slopes),
                              mu = fung_reg_lm_slope_air_volume)


```

plots
```{r}
my_regress_plot <- function(regress_data, wvol_color = "black", detwt_color = "black") {
  gg1 <- ggplot(regress_data, aes(x=log(water_volume), y=log_richness)) + 
  geom_point(color = wvol_color) +
  stat_smooth(method = "lm", 
              formula = y ~ x, 
              geom = "smooth",
              se = F,
              color = wvol_color) +
  #theme(axis.text = element_text(color = "black", size = 14)) +
  labs(y = "log(richness)") +
  theme_classic() +
  theme(axis.text = element_text(color = "black", size = 12),
        axis.title = element_text(color = "black", size = 13)) 
  
  gg2 <- ggplot(regress_data, aes(x=log(detritus_wt), y=log_richness)) + 
  geom_point(color = detwt_color) +
  stat_smooth(method = "lm", 
              formula = y ~ x, 
              geom = "smooth",
              se = F,
              color = detwt_color) +
  # geom_errorbar(aes(ymin=log_rich-se, ymax=log_rich+se),
  #               width=.2,
  #               color = met.brewer("NewKingdom")[4]) +
  labs(y = "log(richness)", x = "log(detritus_weight)") +
  theme_classic() +
  theme(axis.text = element_text(color = "black", size = 12),
        axis.title = element_text(color = "black", size = 13)) 
  
  gglist <- list(gg1, gg2)
  
  return(gglist)
  
  
}


fung_inv_plot <- my_regress_plot(fung_inv_regress_dat, 
                                 wvol_color = met.brewer("Nattier")[2],
                                 detwt_color = met.brewer("Nattier", 4)[2])
fung_inv_plot <- lapply(fung_inv_plot, function(x) x + ggtitle("Fungi, Invertebrates"))

fung_det_plot <- my_regress_plot(fung_det_regress_dat,
                                 detwt_color = met.brewer("Nattier", 4)[3])
fung_det_plot <- lapply(fung_det_plot, function(x) x + ggtitle("Fungi, Detritus"))

fung_water_plot <- my_regress_plot(fung_water_regress_dat,
                                   wvol_color = met.brewer("Nattier", 4)[4])
fung_water_plot <- lapply(fung_water_plot, function(x) x + ggtitle("Fungi, Water"))

# fung_wpool_plot <- my_regress_plot(fung_wpool_regress_dat$summarized_data)
# fung_wpool_plot <- lapply(fung_wpool_plot, function(x) x + ggtitle("Fungi, WPool"))

fung_air_plot <- ggplot(fung_air, aes(x=log(air_volume), y=log_richness)) + 
  geom_point(color = met.brewer("Nattier")[1]) +
  stat_smooth(method = "lm", 
              formula = y ~ x, 
              geom = "smooth",
              se = F,
              color = met.brewer("Nattier")[1]) +
  # geom_errorbar(aes(ymin=log_rich-se, ymax=log_rich+se), 
  #               width=.2,
  #               color = met.brewer("NewKingdom")[1]) +
  labs(y = "log(richness)", title = "Fungi, Air") +
  theme_classic() +
  theme(axis.text = element_text(color = "black", size = 12),
        axis.title = element_text(color = "black", size = 13)) 

fung_plots <- fung_inv_plot[[1]] + fung_inv_plot[[2]] +
  fung_water_plot[[1]] +
  fung_det_plot [[2]] + plot_layout(nrow = 2)

#ggsave("../figures/truly_nonpool_all_fungi_sar_regressions.png", height = 12, width = 9.5)

```


Table
```{r}
# things to put in table = locus, compartment, R2, p-value, slope, t-test p-value, intercept

fung_tbl_detritus_wt <- data.frame(Locus = rep("Fungi", 3),
                       Compartment = c("Invertebrates", "Detritus", "Water (W02 and WPool"),
                       Volume = rep("Detritus_weight_mg", 3))

fung_tbl_detritus_wt$R2 = sapply(fung_regular_lms_detritus_wt,
                     function(x) summary(x)$r.squared)

fung_tbl_detritus_wt$pval = sapply(fung_regular_lms_detritus_wt,
                       function(x) summary(x)$coefficients[,4][2])

fung_tbl_detritus_wt$Slope = sapply(fung_regular_lms_detritus_wt,
                        function(x) coef(x)[2])

fung_tbl_detritus_wt$bootstrap_ttest_pval <- sapply(fung_tts_detritus_wt, 
                                        function(x) x$p.value)

fung_tbl_detritus_wt$Intercept = sapply(fung_regular_lms_detritus_wt,
                        function(x) coef(x)[1])


##### water volume
fung_tbl_water_volume <- data.frame(Locus = rep("Fungi", 3),
                       Compartment = c("Invertebrates", "Detritus", "Water (W02 and WPool)"),
                       Volume = rep("Water_volume_mL", 3))

fung_tbl_water_volume$R2 = sapply(fung_regular_lms_water_volume,
                     function(x) summary(x)$r.squared)

fung_tbl_water_volume$pval = sapply(fung_regular_lms_water_volume,
                       function(x) summary(x)$coefficients[,4][2])

fung_tbl_water_volume$Slope = sapply(fung_regular_lms_water_volume,
                        function(x) coef(x)[2])

fung_tbl_water_volume$bootstrap_ttest_pval <- sapply(fung_tts_water_volume, 
                                        function(x) x$p.value)

fung_tbl_water_volume$Intercept = sapply(fung_regular_lms_water_volume,
                        function(x) coef(x)[1])


fung_full_tbl <- rbind(fung_tbl_detritus_wt, fung_tbl_water_volume)


fung_row_air <- c(Locus = "Fungi",
                  Compartment = "Air",
                  Volume = "Air_volume_L",
                  R2 = summary(fung_regular_lm_air_volume)$r.squared,
                  pval = summary(fung_regular_lm_air_volume)$coefficients[,4][2],
                  Slope = coef(fung_regular_lm_air_volume)[2],
                  bootstrap_ttest_pval = fung_tts_air_volume$p.value,
                  Intercept = coef(fung_regular_lm_air_volume)[1]
                  )

fung_full_tbl <- rbind(fung_full_tbl, fung_row_air)

# take out things not focusing on
fung_final_tbl <- fung_full_tbl[-c(5,3),]

fung_final_tbl <- fung_final_tbl[c(3,1,4,2,5),]

```


# Bacteria

Regressions and bootstraps
```{r}
# volume = detritus weight
bact_regress_data_list = list(bact_inv_regress_dat,
                              bact_det_regress_dat,
                              bact_water_regress_dat)

bact_detwt_boot_slopes <- lapply(bact_regress_data_list,
                                 function(data) bfun(x = log(data$detritus_wt), y = data$log_richness, R = 10000, seed=100))


bact_regular_lms_detritus_wt <- lapply(bact_regress_data_list,
                           function(x) lm(log_richness ~ log(detritus_wt), data = x))

bact_reg_lm_slopes_detritus_wt <- lapply(bact_regular_lms_detritus_wt, function(y) coef(y)[2])

bact_tts_detritus_wt <- list()

for (i in 1:length(bact_regress_data_list)) {
  bact_tts_detritus_wt[[i]] = t.test(unlist(bact_detwt_boot_slopes[[i]]), mu = bact_reg_lm_slopes_detritus_wt[[i]])
}


##### water volume
bact_wvol_boot_slopes <- lapply(bact_regress_data_list,
                                 function(data) bfun(x = log(data$water_volume), y = data$log_richness, R = 10000, seed=88888))

bact_regular_lms_water_volume <- lapply(bact_regress_data_list,
                           function(x) lm(log_richness ~ log(water_volume), data = x))

bact_reg_lm_slopes_water_volume <- lapply(bact_regular_lms_water_volume, function(y) coef(y)[2])

bact_tts_water_volume <- list()

for (i in 1:length(bact_regress_data_list)) {
  bact_tts_water_volume[[i]] = t.test(unlist(bact_wvol_boot_slopes[[i]]), mu = bact_reg_lm_slopes_water_volume[[i]])
}
```

Air
```{r}
# air
# richness per sample vs water volume

# We left the sampler out in the field to collect samples of the air microbiome above the bromeliad garden for 1 hour, 2.5 hours, 4.5 hours, and 6 hours with n=6 replicates per time point. These time points correspond to 990, 2475, 4455, and 5940 liters respectively. 

bact_air <- bact_rep_dat[bact_rep_dat$newCompartment=="Air",]
bact_air <- subset(bact_air, select = -c(detritus_wt, water_volume))
bact_air$combo <- paste(bact_air$Bro1, bact_air$Compartment, sep = "_")

bact_air<- bact_air %>%
  mutate(air_volume = case_when(
    Compartment == "T1" ~ 990,
    Compartment == "T2" ~ 2475,
    Compartment == "T3" ~ 4455,
    Compartment == "T4" ~ 5940
  ))

bact_air$log_richness <- log(bact_air$richness)

# bact_air_summ <- bact_air %>% dplyr::group_by(Compartment) %>% dplyr::summarize(log_rich = mean(log_richness)) # S2 only has replicates 4, 5, 6 and T1 only
# # 6 total reps for T1 though... ask NH whether to combine S1 and S2 for T1
# bact_air_summ_se <- bact_air %>% group_by(Compartment) %>% summarize(se = sd(log_richness)/sqrt(length(log_richness)))
# 
# bact_air_summ$se <- bact_air_summ_se$se
# #bact_air_summ$combo <- paste(bact_air_summ$Bro1, bact_air_summ$Compartment, sep="_")
# 
# bact_air_summ<- bact_air_summ %>%
#   mutate(air_volume = case_when(
#     Compartment == "T1" ~ 990,
#     Compartment == "T2" ~ 2475,
#     Compartment == "T3" ~ 4455,
#     Compartment == "T4" ~ 5940
#   ))


##### bootstrap and regressions
bact_air_boot_slopes <- bfun(x = log(bact_air$air_volume), y = bact_air$log_richness, R = 10000, seed=911)

bact_regular_lm_air_volume <- lm(log_richness ~ log(air_volume), data = bact_air)

bact_reg_lm_slope_air_volume <- coef(bact_regular_lm_air_volume)[2]

bact_tts_air_volume <- t.test(bact_air_boot_slopes,
                              mu = bact_reg_lm_slope_air_volume)


```
plots
```{r}
bact_inv_plot <- my_regress_plot(bact_inv_regress_dat, 
                                 wvol_color = met.brewer("Nattier")[2],
                                 detwt_color = met.brewer("Nattier", 4)[2])
bact_inv_plot <- lapply(bact_inv_plot, function(x) x + ggtitle("Bacteria, Invertebrates"))

bact_det_plot <- my_regress_plot(bact_det_regress_dat,
                                 detwt_color = met.brewer("Nattier", 4)[3])
bact_det_plot <- lapply(bact_det_plot, function(x) x + ggtitle("Bacteria, Detritus"))

bact_water_plot <- my_regress_plot(bact_water_regress_dat,
                                   wvol_color = met.brewer("Nattier", 4)[4])
bact_water_plot <- lapply(bact_water_plot, function(x) x + ggtitle("Bacteria, Water"))

# bact_wpool_plot <- my_regress_plot(bact_wpool_regress_dat$summarized_data)
# bact_wpool_plot <- lapply(bact_wpool_plot, function(x) x + ggtitle("Bacteria, WPool"))

bact_air_plot <- ggplot(bact_air, aes(x=log(air_volume), y=log_richness)) + 
  geom_point(color = met.brewer("Nattier")[1]) +
  stat_smooth(method = "lm", 
              formula = y ~ x, 
              geom = "smooth",
              se = F,
              color = met.brewer("Nattier")[1]) +
  # geom_errorbar(aes(ymin=log_rich-se, ymax=log_rich+se), 
  #               width=.2,
  #               color = met.brewer("NewKingdom")[1]) +
  labs(y = "log(richness)", title = "Bacteria, Air") +
  theme_classic() +
  theme(axis.text = element_text(color = "black", size = 12),
        axis.title = element_text(color = "black", size = 13)) 

bact_plots <- bact_inv_plot[[1]] + bact_inv_plot[[2]] +
  bact_water_plot[[1]] +
  bact_det_plot [[2]] + plot_layout(nrow = 2)


#ggsave("../figures/truly_nonpool_all_bact_sar_regressions.png", bact_plots, height = 12, width = 9.5)
```

```{r}
all_sar_plots <- 
  fung_det_plot [[2]] + fung_water_plot[[1]] + fung_inv_plot[[2]] + fung_inv_plot[[1]] + 
  bact_det_plot [[2]] + bact_water_plot[[1]] + bact_inv_plot[[2]] + bact_inv_plot[[1]] +
  plot_layout(nrow = 2)

ggsave("../figures/for_manuscript/all_sar_nonpool_plots.pdf", width=16, height=7.5)
ggsave("../figures/for_manuscript/all_sar_nonpool_plots.png", width=16, height=7.5)


air_plots <- fung_air_plot + bact_air_plot

ggsave("../figures/for_manuscript/air_sar_plots.png", air_plots, width= 10, height=4.5)
ggsave("../figures/for_manuscript/air_sar_plots.pdf", air_plots, width= 10, height=4.5)
```


Table
```{r}
# things to put in table = locus, compartment, R2, p-value, slope, t-test p-value, intercept

bact_tbl_detritus_wt <- data.frame(Locus = rep("Bacteria", 3),
                       Compartment = c("Invertebrates", "Detritus", "Water (W02)"),
                       Volume = rep("Detritus_weight_mg", 3))

bact_tbl_detritus_wt$R2 = sapply(bact_regular_lms_detritus_wt,
                     function(x) summary(x)$r.squared)

bact_tbl_detritus_wt$pval = sapply(bact_regular_lms_detritus_wt,
                       function(x) summary(x)$coefficients[,4][2])

bact_tbl_detritus_wt$Slope = sapply(bact_regular_lms_detritus_wt,
                        function(x) coef(x)[2])

bact_tbl_detritus_wt$bootstrap_ttest_pval <- sapply(bact_tts_detritus_wt, 
                                        function(x) x$p.value)

bact_tbl_detritus_wt$Intercept = sapply(bact_regular_lms_detritus_wt,
                        function(x) coef(x)[1])


##### water volume
bact_tbl_water_volume <- data.frame(Locus = rep("Bacteria", 3),
                       Compartment = c("Invertebrates", "Detritus", "Water (W02)"),
                       Volume = rep("Water_volume_mL", 3))

bact_tbl_water_volume$R2 = sapply(bact_regular_lms_water_volume,
                     function(x) summary(x)$r.squared)

bact_tbl_water_volume$pval = sapply(bact_regular_lms_water_volume,
                       function(x) summary(x)$coefficients[,4][2])

bact_tbl_water_volume$Slope = sapply(bact_regular_lms_water_volume,
                        function(x) coef(x)[2])

bact_tbl_water_volume$bootstrap_ttest_pval <- sapply(bact_tts_water_volume, 
                                        function(x) x$p.value)

bact_tbl_water_volume$Intercept = sapply(bact_regular_lms_water_volume,
                        function(x) coef(x)[1])


bact_full_tbl <- rbind(bact_tbl_detritus_wt, bact_tbl_water_volume)


bact_row_air <- c(Locus = "Bacteria",
                  Compartment = "Air",
                  Volume = "Air_volume_L",
                  R2 = summary(bact_regular_lm_air_volume)$r.squared,
                  pval = summary(bact_regular_lm_air_volume)$coefficients[,4][2],
                  Slope = coef(bact_regular_lm_air_volume)[2],
                  bootstrap_ttest_pval = bact_tts_air_volume$p.value,
                  Intercept = coef(bact_regular_lm_air_volume)[1]
                  )

bact_full_tbl <- rbind(bact_full_tbl, bact_row_air)

# take out things not focusing on
bact_final_tbl <- bact_full_tbl[-c(5,3),]

bact_final_tbl <- bact_final_tbl[c(3,1,4,2,5),]

```

```{r}
full_full_tbl <- rbind(fung_final_tbl, bact_final_tbl)
full_full_tbl$bootstrap_ttest_pval <- as.numeric(full_full_tbl$bootstrap_ttest_pval)

### correcting P-values for multiple comparisons
library(multtest)  

# Set seed for repeatability
set.seed(2024)
  
adj_p_values = mt.rawp2adjp(full_full_tbl$bootstrap_ttest_pval)
    
adj_p_mat = adj_p_values$adjp[, c("rawp", "BH")]
  
adj_p_ordered = adj_p_mat[,2][order(adj_p_values$index)]

full_full_tbl$bootstrap_ttest_pval <- adj_p_ordered

#saveRDS(full_full_tbl, "../intermediates/truly_nonpooled_regression_stats_table.rds")
```


# z-value comparisons among compartments, within and between loci
Bootstrap to generate variance on the observed data slope, repeated 1,000 times
```{r}

# bootstrap function
bfun <- \(x, y, R=1000, seed=134) {
  FUN <- \() {
    i <- sample(seq_len(length(x)), replace=TRUE)
    y <- y[i]
    x <- x[i]
    return(lm(y ~ x)$coefficients)

    #return()
  }
  set.seed(seed)
  r <- t(replicate(R, FUN()))
  bootdist <- r[, 2]

  return(bootdist)

}

# get_ci <- function(slope_distribution_vector) {
#   sample.mean <- mean(slope_distribution_vector)
#   sample.n <- length(slope_distribution_vector)
#   sample.sd <- sd(slope_distribution_vector)
#   sample.se <- sample.sd/sqrt(sample.n)
#   
#   alpha = 0.05
#   degrees.freedom = sample.n - 1
#   t.score = qt(p=alpha/2, df=degrees.freedom,lower.tail=F)
#   
#   margin.error <- t.score * sample.se
#   
#   lower.bound <- sample.mean - margin.error
#   upper.bound <- sample.mean + margin.error
#   
#   return(paste(round(lower.bound, 4),
#                round(upper.bound, 4),
#                sep=","))
# 
# }

```

```{r}
fung_dw <- lapply(fung_regress_data_list,
                  function(x) bfun(y = x$log_richness, x = log(x$detritus_wt), R = 1000, seed=4))

#fung_dw_ci <- lapply(fung_dw, get_ci)

fung_wv <- lapply(fung_regress_data_list,
                  function(x) bfun(y = x$log_richness, x = log(x$water_volume), R = 1000, seed=12))

#fung_wv_ci <- lapply(fung_wv, get_ci)

fung_a <- bfun(y = fung_air$log_richness, x = log(fung_air$air_volume), R = 1000, seed=13)
fung_a <- fung_a[!is.na(fung_a)]

#fung_a_ci <- get_ci(fung_a)


bact_dw <- lapply(bact_regress_data_list,
                  function(x) bfun(y = x$log_richness, x = log(x$detritus_wt), R=1000, seed=81))

#bact_dw_ci <- lapply(bact_dw, get_ci)

bact_wv <- lapply(bact_regress_data_list ,
                  function(x) bfun(y = x$log_richness, x = log(x$water_volume), R=1000, seed=99))

#bact_wv_ci <- lapply(bact_wv, get_ci)

bact_a <- bfun(y = bact_air$log_richness, x = log(bact_air$air_volume), R=1000, seed=8)
bact_a <- bact_a[!is.na(bact_a)]

#bact_a_ci <- get_ci(bact_a)


# append to full table
ci_vector <- c(fung_wv_ci[[1]],
                    fung_dw_ci[[1]],
                    fung_wv_ci[[3]],
                    fung_dw_ci[[2]],
                    fung_a_ci,
                    bact_wv_ci[[1]],
                    bact_dw_ci[[1]],
                    bact_wv_ci[[3]],
                    bact_dw_ci[[2]],
                    bact_a_ci
                    )

new_full_table <- full_full_tbl
new_full_table$CI_95 <- ci_vector

new_full_table <- new_full_table %>% dplyr::select(Locus, Compartment, Volume, R2, pval, Slope, CI_95, bootstrap_ttest_pval, Intercept)
#saveRDS(new_full_table, "../intermediates/truly_nonpool_regression_stats_with_bootstrapped_confint.rds")

#write.csv(new_full_table, "../outputs/truly_nonpool_regression_stats_with_bootstrapped_confint.csv")
```

Pairwise t-tests
```{r}
# within locus, across compartments
library(rstatix)

long_dat1 <- data.frame(type = c(rep("fung_inverts_dw", length(fung_dw[[1]])),
                                 rep("fung_detritus_dw", length(fung_dw[[2]])),
                                 #rep("fung_water_dw", length(fung_dw[[3]])),
                                 rep("fung_inverts_wv", length(fung_wv[[1]])),
                                 #rep("fung_detritus_wv", length(fung_wv[[2]])),
                                 rep("fung_water_wv", length(fung_wv[[3]]))
                                # rep("fung_air", length(fung_a)) 
                                 ),
                            slope = c(fung_dw[[1]],
                                  fung_dw[[2]],
                                  #fung_dw[[3]],
                                  fung_wv[[1]],
                                  #fung_wv[[2]],
                                  fung_wv[[3]])
                                 # fung_a)
)

fung_within_locus <- rstatix::pairwise_t_test(long_dat1, slope ~ type, p.adjust.method = "BH", pool.sd=FALSE, detailed = TRUE)

long_dat2 <- data.frame(type = c(rep("bact_inverts_dw", length(bact_dw[[1]])),
                                 rep("bact_detritus_dw", length(bact_dw[[2]])),
                                 #rep("bact_water_dw", length(bact_dw[[3]])),
                                 rep("bact_inverts_wv", length(bact_wv[[1]])),
                                 #rep("bact_detritus_wv", length(bact_wv[[2]])),
                                 rep("bact_water_wv", length(bact_wv[[3]]))
                                 #rep("bact_air", length(bact_a)) 
                                 ),
                            slope = c(bact_dw[[1]],
                                  bact_dw[[2]],
                                  #bact_dw[[3]],
                                  bact_wv[[1]],
                                  #bact_wv[[2]],
                                  bact_wv[[3]])
                                 # bact_a)
)

bact_within_locus <- rstatix::pairwise_t_test(long_dat2, slope ~ type, p.adjust.method = "BH", pool.sd=FALSE, detailed = TRUE)

## prep within locus
all_within_locus <- rbind(fung_within_locus, bact_within_locus)
all_within_locus <- select(all_within_locus, group1, group2, statistic, df, p.adj)

all_within_locus_renamed <- all_within_locus

#all_within_locus_renamed[all_within_locus_renamed=="fung_air"] <- "Fungi, Air"
all_within_locus_renamed[all_within_locus_renamed=="fung_detritus_dw"] <- "Fungi, Detritus"
all_within_locus_renamed[all_within_locus_renamed=="fung_inverts_dw"] <- "Fungi, Invertebrates (Detritus weight)"
all_within_locus_renamed[all_within_locus_renamed=="fung_inverts_wv"] <- "Fungi, Invertebrates (Water volume)"
all_within_locus_renamed[all_within_locus_renamed=="fung_water_wv"] <- "Fungi, Water"

#all_within_locus_renamed[all_within_locus_renamed=="bact_air"] <- "Bacteria, Air"
all_within_locus_renamed[all_within_locus_renamed=="bact_detritus_dw"] <- "Bacteria, Detritus"
all_within_locus_renamed[all_within_locus_renamed=="bact_inverts_dw"] <- "Bacteria, Invertebrates (Detritus weight)"
all_within_locus_renamed[all_within_locus_renamed=="bact_inverts_wv"] <- "Bacteria, Invertebrates (Water volume)"
all_within_locus_renamed[all_within_locus_renamed=="bact_water_wv"] <- "Bacteria, Water"

all_within_locus_renamed$p.adj <- round(all_within_locus_renamed$p.adj, digits = 3)
#all_within_locus_renamed <- select(all_within_locus_renamed, group1, group2, p.adj)

write.csv(all_within_locus_renamed, "../outputs/for_manuscript/bootslope_ttest_within_locus_noair_morestats.csv")


# inv, det, water, air
long_data <- data.frame(type = c(rep("fung_inverts_dw", length(fung_dw[[1]])),
                                 rep("fung_detritus_dw", length(fung_dw[[2]])),
                                 #rep("fung_water_dw", length(fung_dw[[3]])),
                                 rep("fung_inverts_wv", length(fung_wv[[1]])),
                                 #rep("fung_detritus_wv", length(fung_wv[[2]])),
                                 rep("fung_water_wv", length(fung_wv[[3]])),
                                 #rep("fung_air", length(fung_a)),
                                 
                                 rep("bact_inverts_dw", length(bact_dw[[1]])),
                                 rep("bact_detritus_dw", length(bact_dw[[2]])),
                                 #rep("bact_water_dw", length(bact_dw[[3]])),
                                 rep("bact_inverts_wv", length(bact_wv[[1]])),
                                 #rep("bact_detritus_wv", length(bact_wv[[2]])),
                                 rep("bact_water_wv", length(bact_wv[[3]]))
                                # rep("bact_air", length(bact_a))
                                 
                                 ),
                        slope = c(fung_dw[[1]],
                                  fung_dw[[2]],
                                  #fung_dw[[3]],
                                  fung_wv[[1]],
                                  #fung_wv[[2]],
                                  fung_wv[[3]],
                                  #fung_a,
                                  
                                  bact_dw[[1]],
                                  bact_dw[[2]],
                                  #bact_dw[[3]],
                                  bact_wv[[1]],
                                  #bact_wv[[2]],
                                  bact_wv[[3]]
                                  #bact_a
                                  ))

#between_locus <- rstatix::pairwise_t_test(long_data, slope ~ type, p.adjust.method = "BH")

btwn_inv_wv <- t.test(fung_wv[[1]], bact_wv[[1]])
btwn_inv_dw <- t.test(fung_dw[[1]], bact_dw[[1]])
btwn_det_dw <- t.test(fung_dw[[2]], bact_dw[[2]])
btwn_wat_wv <- t.test(fung_wv[[3]], bact_wv[[3]])
btwn_air <- t.test(fung_a, bact_a)

btwn_list <- list(btwn_inv_wv, btwn_inv_dw, btwn_det_dw, btwn_wat_wv) # no air

new_btwn_list <- list()

for (i in 1:length(btwn_list)) {
  mydat = data.frame(group = btwn_list[[i]]$data.name,
                     statistic = btwn_list[[i]]$statistic,
                     df = btwn_list[[i]]$parameter,
                     pval = btwn_list[[i]]$p.value
                     )
  
  new_btwn_list[[i]] <- mydat
  
}

full_btwn <- do.call("rbind", new_btwn_list)

### correcting P-values for multiple comparisons
library(multtest)  

# Set seed for repeatability
set.seed(202415)
  
adj_p_values = mt.rawp2adjp(full_btwn$pval)
    
adj_p_mat = adj_p_values$adjp[, c("rawp", "BH")]
  
adj_p_ordered = adj_p_mat[,2][order(adj_p_values$index)]

full_btwn$adjusted_pval <- adj_p_ordered


full_btwn_renamed <- full_btwn

full_btwn_renamed$group2 <- full_btwn$group
full_btwn_renamed <- full_btwn_renamed %>% select(group, group2, statistic, df, adjusted_pval)

full_btwn_renamed$group <- c("Fungi, Invertebrates (Water volume)",
                             "Fungi, Invertebrates (Detritus weight)",
                             "Fungi, Detritus",
                             "Fungi, Water")

full_btwn_renamed$group2 <- c("Bacteria, Invertebrates (Water volume)",
                             "Bacteria, Invertebrates (Detritus weight)",
                             "Bacteria, Detritus",
                             "Bacteria, Water")

full_btwn_renamed$adjusted_pval <- round(full_btwn_renamed$adjusted_pval, digits = 3)

write.csv(full_btwn_renamed, "../outputs/for_manuscript/bootslope_ttest_btwn_loci_noair_morestats.csv")

# pair_t <- pairwise.t.test(long_data$slope, long_data$type, p.adjust.method = "BH")[[3]]
# pair_t2 <- apply(pair_t, 2, function(x) ifelse(!is.na(x), round(x, 7), NA))

# saveRDS(pair_t2, "../intermediates/truly_nonpool_pairwise_ttest_boot_slopes_between_compartments.rds")
# write.csv(pair_t2, "../outputs/truly_nonpool_pairwise_ttest_boot_slopes_between_compartments.csv")
```

## save out data for slope comparison
```{r}
# saveRDS(fung_regress_data_list, "../intermediates/truly_nonpooled_fung_regress_data_list.rds")
# saveRDS(fung_air, "../intermediates/truly_nonpooled_fung_air_data.rds")
# 
# saveRDS(bact_regress_data_list, "../intermediates/truly_nonpooled_bact_regress_data_list.rds")
# saveRDS(bact_air, "../intermediates/truly_nonpooled_bact_air_data.rds")

# read out
fung_regress_data_list <- readRDS("../intermediates/truly_nonpooled_fung_regress_data_list.rds")
bact_regress_data_list <- readRDS("../intermediates/truly_nonpooled_bact_regress_data_list.rds")

fung_air <- readRDS("../intermediates/truly_nonpooled_fung_air_data.rds")
bact_air <- readRDS("../intermediates/truly_nonpooled_bact_air_data.rds")
```

### z-score plot

```{r}
# order of list - inv, det, water
# fung_regress_data_list = list(fung_inv_regress_dat,
#                               fung_det_regress_dat,
#                               fung_water_regress_dat)

fung_detwt = sapply(fung_regular_lms_detritus_wt,
                        function(x) coef(x)[2])

fung_wvol = sapply(fung_regular_lms_water_volume,
                        function(x) coef(x)[2])

bact_detwt = sapply(bact_regular_lms_detritus_wt,
                        function(x) coef(x)[2])

bact_wvol = sapply(bact_regular_lms_water_volume,
                        function(x) coef(x)[2])


slope_df <- data.frame(locus = c(rep("Fungi", 4),
                                 rep("Bacteria", 4)),
                       compartment = c("Detritus", "Water", "Invertebrates (d)", "Invertebrates (w)",
                                       "Detritus", "Water", "Invertebrates (d)", "Invertebrates (w)"),
                       slope = c(fung_detwt[2],
                                 fung_wvol[3],
                                 fung_detwt[1],
                                 fung_wvol[1],
                                 bact_detwt[2],
                                 bact_wvol[3],
                                 bact_detwt[1],
                                 bact_wvol[1]
                         
                       )
)

xx = factor(unique(slope_df$locus), levels = c("Fungi", "Bacteria"))
yy = factor(unique(slope_df$compartment), levels = c("Detritus", "Water", "Invertebrates (d)", "Invertebrates (w)"))

z_plot <- ggplot(slope_df, aes(fill=factor(compartment, levels = c("Detritus", "Water", "Invertebrates (d)", "Invertebrates (w)")), 
                                y=slope, 
                                x=factor(locus, levels = c("Fungi", "Bacteria")))) + 
    geom_bar(position="dodge", stat="identity") +
  scale_fill_manual(values = met.brewer("Cassatt1", 5)) +
  theme_minimal() +
     # theme(panel.grid.major=element_blank(), 
    #panel.grid.minor=element_blank(), 
    #panel.background=element_blank(), 
    #axis.line = element_line(colour="black"),
   theme(axis.text = element_text(color = "black", size=12),
         legend.text = element_text(size = 13),
         axis.title = element_text (size = 14),
         legend.title = element_text(size = 14)) +
  labs(x = "Microbe", y = "z-score", fill = "Compartment") +
  scale_y_continuous(breaks = c(-0.15, -0.1, -0.05, 0, 0.05, 0.1, 0.15),limits = c(c(-0.15, 0.15)))

ggsave("../figures/for_manuscript/z_score_plot2.pdf", width=8, height=5.5)
ggsave("../figures/for_manuscript/z_score_plot2.png", width=8, height=5.5)


# only significant SARs
sig_sar_df <- slope_df[c(2,5,6),]
z_plot_sig <- ggplot(sig_sar_df, aes(fill=factor(compartment, levels = c("Detritus", "Water")), 
                                y=slope, 
                                x=factor(locus, levels = c("Fungi", "Bacteria")))) + 
    geom_bar(position = position_dodge2(width = 0.9, preserve = "single"), stat="identity") +
  scale_fill_manual(values = met.brewer("Cassatt1", 5)) +
    theme(panel.grid.major=element_blank(),
    panel.grid.minor=element_blank(),
    panel.background=element_blank(),
    axis.line = element_line(colour="black")) +
   theme_minimal() +
     # theme(panel.grid.major=element_blank(), 
    #panel.grid.minor=element_blank(), 
    #panel.background=element_blank(), 
    #axis.line = element_line(colour="black"),
   theme(axis.text = element_text(color = "black", size=12),
         legend.text = element_text(size = 13),
         axis.title = element_text (size = 14),
         legend.title = element_text(size = 14)) +
  labs(x = "Microbe", y = "z-score", fill = "Compartment") +
  scale_y_continuous(breaks = c(-0.1,0,0.1), limits = c(-0.113, 0.111), expand = c(0, 0))

ggsave("../figures/for_manuscript/z_score_plot3.pdf", width=8, height=5.5)
ggsave("../figures/for_manuscript/z_score_plot3.png", width=8, height=5.5)
```
